+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- [TD-4134] `taxonomy_roles` notifications
+
 ## [4.29.0] 2021-10-04
 
 ### Added